requirements.txt
discord.py
discord.py[voice]
python-dotenv
PyNaCl
yt_dlp
ffmpeg